fix error when delete post with empty files

+++ src/main/java/com/ttree/ttree/service/FinalPTFileService.java
@@ -40,7 +40,11 @@ public class FinalPTFileService {
 
     @Transactional
     public void deleteFinalPTFile(Long id){
-        finalPTFileRepository.deleteById(id);
+        if(finalPTFileRepository.findById(id).isPresent()) {
+            finalPTFileRepository.deleteById(id);
+        }else{
+            System.out.println("file is already empty");
+        }
     }
 
 }
